Iteration 2 of Odin-Recipes
Create a new directory within the odin-recipes directory and name it recipes.
Create a new HTML file within the recipes directory and name it after
the recipe it will contain. For example lasagna.html.
h1 heading with the recipe’s name as its content.
In the index.html file, add a link to the recipe page.
The text of the link is the recipe name.