Iteration 3: Recipe Page Content
The recipe page should have the following content:
An image of the finished dish under the h1 heading that you added earlier.
A “Description” heading followed by a paragraph or two describing the recipe.
An “Ingredients” heading followed by an unordered list of the ingredients
A “Steps” heading followed by an ordered list of the steps needed for making the dish.